Fixed most warnings (Not fixed: unused parameters when the parameter is required by a superclass interface.)
git-svn-id: https://wxguitest.svn.sourceforge.net/svnroot/wxguitest/trunk@59 64288482-8357-404e-ad65-de92a562ee98

+++ current/src/swWxGuiTesting/CapturedEvents/swCRTextUpdateEvent.cpp
@@ -179,7 +179,7 @@ wxString CRTextUpdateEvent::GetEscaped (const wxString &str) const
 {
     wxString escaped;
 
-    for (int i = 0; i < str.Length (); i++) {
+    for (size_t i = 0; i < str.Length (); i++) {
 
         if (str[i] == 92) { // '\'
